Add files via upload
Three Different codes working with the servo. One is a little more complicated, and I did not understand it as much, and one does not use Interrupts.  All should work fine, and hopefully can be implemented into the overall code.  

All three use the built in buttons on the mps P1.1, and P1.2, and the two with interrupts use P1.3 as the signal to the servo, while the code without interrupts use P1.6 as the signal pin.